updated pvwatts code that uses the various RECORE library facilities to divide up the calculation work
For the configurations tested against the original C, day at a time, code, the results are exceptionally close.

The very minor differences are due to the fact that new version's cell temperature calculation keeps track of the
module temperature from day 1, while the original resets the cell temperature calculation each day (does 1 day at a
time calculations).  

If the local variables in 'celltemp(..)' in pvwatts1.c are declared as static (so that their values are saved 
between calls (as was intended in the original fortran implementation in PVFORM), the results match exactly.

This is the explanation of slight differences between the current (as of apr 2012) online PVWatts V1 and the 
'pvwattsv1' implementation in SSC.  For most systems the difference will not be noticable.