PV: added error if temp or WS not in weather file
Added an error to check if the temperature or WS columns are missing
from the weather file. If the column is there but the cells are blank,
SAM will read them in as zeros, so not a perfect check, but it will
catch the columns being completely missing from the weather file,
instead of giving negative answers. Fixes #18