Fix warning: statement has no effect [-Wunused-value].
There are many instances in this patch of taking the address of a
variable and casting the pointer to void * in the hope of silencing
compiler warnings about the variable being unused. It would be far
better to find out why the variable isn't used and fixing the root
cause than to mask the problem with void * casts.