Added POA decomposition warnings
Added warnings when the final DNI or DHI calculated by the POA decomposition algorithm is negative. In order to do this, I had to get rid of some of the intermediate checks for negative irradiance. This changes the behavior of the decomposition model slightly, but only when the model is calculating negative irradiance to begin with. Overall it has a very small effect on simulations. Fixes #208