Updated pvsamv1 test for POA decomposition model fix
The changes to the POA decomposition algorithm to allow warnings to be reported caused slight changes in the calculated DNI and calculated DHI reported by the POA decomposition model, primarily in the morning and evening. These slight changes do not affect the annual energy calculations when POA is used directly, but do affect annual energy calculations when the decomped DNI and DHI are used, as is the case for the "POA from pyranometer" irradiance option is used. Therefore, updated the Sky Diffuse model test to match the new results for the POA from pyranometer option.