Eliminate resetting of defocus on first ncall
The TCS kernel will compare the next defocus it is sent to the first defocus it sent to decide if the module should be run again. By resetting the defocus to 1, if the next value sent is the same as the original, but different than this reset value and all the values dependent on it, than it will be assumed the module has converged even if it hasnt.