DC voltage variable labels #146
"Inverter MPPT 1 Nominal DC voltage" is the nominal operating voltage of the subarray connected to MPPT 1 input.
"Subarray 1 Operating DC voltage" is the operating voltage of Subarray 1.
These variables need to be reworked in pvsamv1 so that the nominal voltage is associated with the subarray and the adjusted operating voltage is associated with the inverter MPPT input.