Updates to fix sunrise/sunset bug at very high latitudes
Updated previous fix to be a more robust fix for sunrise and sunset calculations at high latitudes. This includes calculating sunrise and sunset correctly and allowing power to be calculated when the sunset occurs on the next day (e.g. sunset is at 1:30AM from the previous day, sunrise is at 2:30AM that day). Added a test for the high latitude location that proved problematic. In this update, also changed the nighttime sun angles that are reported so that they represent real values calculated at the actual time in question, and updated the NightCase irradproc tests accordingly (previously, the angles were either set to -999 or leftover values from calculating at noon that day).