Add weather file header data to outputs Fix #453
Add latitude, longitude, elevation, time zone to outputs for pvsamv1.

Change labels and units for lat, lon, elev, tz outputs for pvwattsv8 to be consistent with pvsamv1.

Add latitude, longitude, elevation, year to outputs for windpower. Note that windpower does not read time zone from weather file, so not reporting as an output. Also lat, lon, year do not appear to be used in calculations, but are read in, so I decided to include them in the outputs for completeness.

Use "Location" group name for all.